updated and test
