+++ src/comon/SliderAdv.js
@@ -1,10 +1,8 @@
 import React, { Component } from 'react';
-import RealEstateInfo from '../component/RealEstateInfo';
-import PropertiesContainer from '../component/PropertiesContainer';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import '../component/container.css'
-import { Container, Grid, Segment } from 'semantic-ui-react'
+import { Container } from 'semantic-ui-react'
 import { Carousel } from 'react-bootstrap';
+import './comon.css'
 
 
 class SlideAdv extends Component {
