+++ src/comon/comon.css
@@ -6,3 +6,15 @@
    margin-top: 0;
 }
 
+.dimmer{
+   background-color: dimgray;
+   opacity: 0.7;
+   z-index: 1;
+ }
+ h1.sildeCon.title{
+   color: white;
+   font-weight: bold;
+   z-index: 100;
+ }
+         
+
